+++ boringline/src/main/java/com/casual/home/service/HomeService.java
@@ -7,7 +7,7 @@ import com.casual.home.entity.MessageBean;
 public interface HomeService {
 	public MessageBean insert(MessageBean messageBean);
 
-	public MessageBean selectOne(Integer message_id);
+	public MessageBean selectOne(Long message_id);
 
 	public List<MessageBean> selectAll();
